Flip when kd-tree increments/decrements counts to get a more accurate heuristic.
The previous ordering of these statements, taken from pbrt, disagreed with how the
shapes were assigned to left and right partitions after the fact. Since we assign
a shape whose bound is colocated with the splitting plane to both sides of that
plane, we must order these statements to reflect that.

As a downside, we can now never have empty partitions.

TODO: Figure out how pbrt is partitioning these such that they can get empty partitions.