Futher improve kd-tree build heuristics.
Flip the start/end increments and update the left/right partitioning comparisons
to match. This changes the semantics of the tree slightly: objects with a bound
exactly in-plane are usually only assigned to one of the two halves. (There will
always be at least one such object, since partitions are computed from object
bounds exclusively.) If an object is exactly in-plane, i.e. has zero extent along
the given dimension, a special case will add it to both halves since it would
otherwise be dropped. Assertions were added to verify every object is handled.