Add cube map from inside texture.
Most graphics applications of cubemaps seem to be for skyboxes. This texture
labels the faces as you'd expect to see them from inside the cube, as if the
texture was wallpapered onto it from the inside.

Note that this is NOT the mapping scheme we use. We currently use (from the
previous commit) a scheme where -Z is in the middle and the texture is wrapped
around the outside of the box.

These two approaches differ just in some inverted terms (n -> 1 - n), but the
current setup is for the latter because that seems more useful to us.