Flip signs in sphere intersection to properly complete the right-handed coordinate system transition.
Took a while to get here and I'm having trouble keeping it all in my head, but the basic idea
is this: we have use a right-handed coordinate system, but pbrt uses a left-handed one. This
means we often have to swap Y and Z in equations. Furthermore, we use +Y to mean up, whereas a lot
of graphics/math code you'll find uses +Z to mean up. We follow the example of pbrt pg. 57 which
shows Y pointing up and IIRC (though this should be verified) that's what pbrt does for it's
left-handed system as well.

In any case, the point is that I had to triangulate the proper algorithm between pbrt, online
calculators and online math tutorials.

The immediate issue was that the normal was being calculated at something nonsensical.

The root issue was that, while the u/v were being calculated correctly for our coordinate
system (as were the phi/theta derived from them), the calculation from phi/theta back into
Cartesian coordinates was incorrect. By closely examining the calculated values for intersections
in useful places (such as <-1, 0, 0> in the sphere's local coordinate system) and comparing both
the calculated normal to the expected normal (which is the intersection point itself, normalized)
_and_ plotting the u/v axes to ensure they are sane, it was determined that (1) the u/v axes were
not sane and (2) the magnitudes were right but the signs were all wonky.

The fix here makes sense in the context of the original commit(s) that added the right-handed
coordinate system, namely, that a lot of signs had to be flipped. The sphere's transition was
implemented in 328dea11ecd4abbb09f91d161427633628247074, which in retrospect of _course_
didn't work because I only flipped the way the u/v wound around the sphere but I didn't change
the way those u/v were translated into the new coordinate system.